pyproject.toml
[tool.poetry]
name = "demo-ia"
version = "0.1.0"
description = ""
authors = ["Ali Riahi <a.riahii@outlook.fr>"]
readme = "README.md"

packages = [
    { include = "demo_ia" }
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pillow = "^11.3.0"
numpy = "^2.3.2"
tqdm = "^4.67.1"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

